+++ ADP/Fusion/PointL/Term/Epsilon.hs
@@ -16,34 +16,35 @@ import           ADP.Fusion.PointL.Core
 
 
 
-type instance LeftPosTy (IStatic d) Epsilon (PointL I) = IStatic d
+type instance LeftPosTy (IStatic d) (Epsilon Global) (PointL I) = IStatic d
+type instance LeftPosTy (IStatic d) (Epsilon Local) (PointL I) = IVariable d  -- to actually allow local epsilons to work, IStatic does additional static controls
 --type instance LeftPosTy (IVariable d) Epsilon (PointL I) = IVariable d
 
-type instance LeftPosTy (OStatic d) Epsilon (PointL O) = OStatic d
+type instance LeftPosTy (OStatic d) (Epsilon Global) (PointL O) = OStatic d
 
 instance
-  forall pos posLeft m ls i
-  . ( TermStream m (Z:.pos) (TermSymbol M Epsilon) (Elm (Term1 (Elm ls (PointL i))) (Z :. PointL i)) (Z:.PointL i)
-    , posLeft ~ LeftPosTy pos Epsilon (PointL i)
-    , TermStaticVar pos Epsilon (PointL i)
+  forall pos posLeft m ls i lg
+  . ( TermStream m (Z:.pos) (TermSymbol M (Epsilon lg)) (Elm (Term1 (Elm ls (PointL i))) (Z :. PointL i)) (Z:.PointL i)
+    , posLeft ~ LeftPosTy pos (Epsilon lg) (PointL i)
+    , TermStaticVar pos (Epsilon lg) (PointL i)
     , MkStream m posLeft ls (PointL i)
     )
-  ⇒ MkStream m pos (ls :!: Epsilon) (PointL i) where
+  ⇒ MkStream m pos (ls :!: (Epsilon lg)) (PointL i) where
   mkStream Proxy (ls :!: Epsilon) grd us is
     = S.map (\(ss,ee,ii) -> ElmEpsilon ii ss)
-    . addTermStream1 (Proxy ∷ Proxy pos) Epsilon us is
+    . addTermStream1 (Proxy ∷ Proxy pos) (Epsilon @lg) us is
     $ mkStream (Proxy ∷ Proxy posLeft)
                ls
-               (termStaticCheck (Proxy ∷ Proxy pos) Epsilon us is grd)
+               (termStaticCheck (Proxy ∷ Proxy pos) (Epsilon @lg) us is grd)
                us
-               (termStreamIndex (Proxy ∷ Proxy pos) Epsilon is)
+               (termStreamIndex (Proxy ∷ Proxy pos) (Epsilon @lg) is)
   {-# Inline mkStream #-}
 
 
 instance
   ( TermStreamContext m ps ts s x0 i0 is (PointL I)
   )
-  ⇒ TermStream m (ps:.IStatic d) (TermSymbol ts Epsilon) s (is:.PointL I) where
+  ⇒ TermStream m (ps:.IStatic d) (TermSymbol ts (Epsilon lg)) s (is:.PointL I) where
   termStream Proxy (ts:|Epsilon) (us:..LtPointL u) (is:.PointL i)
     = S.map (\(TState s ii ee) ->
               let RiPlI k = getIndex (getIdx s) (Proxy :: PRI is (PointL I))
@@ -66,7 +67,7 @@ instance
 
 instance
   ( TermStreamContext m ps ts s x0 i0 is (PointL O)
-  ) => TermStream m (ps:.OStatic d) (TermSymbol ts Epsilon) s (is:.PointL O) where
+  ) => TermStream m (ps:.OStatic d) (TermSymbol ts (Epsilon lg)) s (is:.PointL O) where
   termStream Proxy (ts:|Epsilon) (us:..LtPointL u) (is:.PointL i)
     = S.map (\(TState s ii ee) ->
                 let io = getIndex (getIdx s) (Proxy :: PRI is (PointL O))
@@ -78,13 +79,20 @@ instance
 -- from @- / Deletion@) on a tape. Hence The instance is only active in
 -- @IStatic 0@ cases.
 
-instance TermStaticVar (IStatic 0) Epsilon (PointL I) where
+instance TermStaticVar (IStatic 0) (Epsilon Global) (PointL I) where
   termStreamIndex Proxy Epsilon (PointL i     ) = PointL i
   termStaticCheck Proxy Epsilon _ (PointL (I# i)) grd = (i ==# 0#) `andI#` grd
+  {-# Inline [0] termStreamIndex #-}
+  {-# Inline [0] termStaticCheck #-}
+
+instance TermStaticVar (IStatic 0) (Epsilon Local) (PointL I) where
+  termStreamIndex Proxy Epsilon (PointL i     ) = PointL i
+  -- | Local epsilons are *always* possible.
+  termStaticCheck Proxy Epsilon _ (PointL (I# i)) grd = grd
   {-# Inline termStreamIndex #-}
   {-# Inline termStaticCheck #-}
 
-instance TermStaticVar (OStatic 0) Epsilon (PointL O) where
+instance TermStaticVar (OStatic 0) (Epsilon Global) (PointL O) where
   termStreamIndex Proxy Epsilon (PointL i     ) = PointL i
   -- |
   --
@@ -100,6 +108,12 @@ instance TermStaticVar (OStatic 0) Epsilon (PointL O) where
   -- TODO But we should probably statically assert that epsilon is the only
   -- symbol on the r.h.s. of whatever we write ...
   termStaticCheck Proxy Epsilon (LtPointL (I# u)) (PointL (I# i)) grd = (u ==# i) `andI#` grd
+  {-# Inline [0] termStreamIndex #-}
+  {-# Inline [0] termStaticCheck #-}
+
+instance TermStaticVar (OStatic 0) (Epsilon Local) (PointL O) where
+  termStreamIndex Proxy Epsilon (PointL i     ) = PointL i
+  termStaticCheck Proxy Epsilon (LtPointL (I# u)) (PointL (I# i)) grd = grd
   {-# Inline [0] termStreamIndex #-}
   {-# Inline [0] termStaticCheck #-}