+++ ADP/Fusion/PointR/Term/Epsilon.hs
@@ -16,31 +16,31 @@ import           ADP.Fusion.PointR.Core
 
 
 
-type instance LeftPosTy (IStatic d) Epsilon (PointR I) = IStatic d
+type instance LeftPosTy (IStatic d) (Epsilon Global) (PointR I) = IStatic d
 
 instance
-  forall pos posLeft m ls i
-  . ( TermStream m (Z:.pos) (TermSymbol M Epsilon) (Elm (Term1 (Elm ls (PointR i))) (Z :. PointR i)) (Z:.PointR i)
-    , posLeft ~ LeftPosTy pos Epsilon (PointR i)
-    , TermStaticVar pos Epsilon (PointR i)
+  forall pos posLeft m ls i lg
+  . ( TermStream m (Z:.pos) (TermSymbol M (Epsilon lg)) (Elm (Term1 (Elm ls (PointR i))) (Z :. PointR i)) (Z:.PointR i)
+    , posLeft ~ LeftPosTy pos (Epsilon lg) (PointR i)
+    , TermStaticVar pos (Epsilon lg) (PointR i)
     , MkStream m posLeft ls (PointR i)
     )
-  ⇒ MkStream m pos (ls :!: Epsilon) (PointR i) where
+  ⇒ MkStream m pos (ls :!: Epsilon lg) (PointR i) where
   mkStream Proxy (ls :!: Epsilon) grd us is
     = S.map (\(ss,ee,ii) -> ElmEpsilon ii ss)
-    . addTermStream1 (Proxy ∷ Proxy pos) Epsilon us is
+    . addTermStream1 (Proxy ∷ Proxy pos) (Epsilon @lg) us is
     $ mkStream (Proxy ∷ Proxy posLeft)
                ls
-               (termStaticCheck (Proxy ∷ Proxy pos) Epsilon us is grd)
+               (termStaticCheck (Proxy ∷ Proxy pos) (Epsilon @lg) us is grd)
                us
-               (termStreamIndex (Proxy ∷ Proxy pos) Epsilon is)
+               (termStreamIndex (Proxy ∷ Proxy pos) (Epsilon @lg) is)
   {-# Inline mkStream #-}
 
 
 instance
   ( TermStreamContext m ps ts s x0 i0 is (PointR I)
   )
-  ⇒ TermStream m (ps:.IStatic d) (TermSymbol ts Epsilon) s (is:.PointR I) where
+  ⇒ TermStream m (ps:.IStatic d) (TermSymbol ts (Epsilon lg)) s (is:.PointR I) where
   termStream Proxy (ts:|Epsilon) (us:..LtPointR u) (is:.PointR i)
     = S.map (\(TState s ii ee) ->
               let RiPrI k = getIndex (getIdx s) (Proxy :: PRI is (PointR I))
@@ -52,9 +52,15 @@ instance
 
 -- TODO need upper bound in @termStaticCheck@ to be able to check against that!
 
-instance TermStaticVar (IStatic 0) Epsilon (PointR I) where
+instance TermStaticVar (IStatic 0) (Epsilon Global) (PointR I) where
   termStreamIndex Proxy Epsilon (PointR i     ) = PointR i
   termStaticCheck Proxy Epsilon (LtPointR (I# u)) (PointR (I# i)) grd = (i ==# u) `andI#` grd
+  {-# Inline termStreamIndex #-}
+  {-# Inline termStaticCheck #-}
+
+instance TermStaticVar (IStatic 0) (Epsilon Local) (PointR I) where
+  termStreamIndex Proxy Epsilon (PointR i     ) = PointR i
+  termStaticCheck Proxy Epsilon (LtPointR (I# u)) (PointR (I# i)) grd = grd
   {-# Inline termStreamIndex #-}
   {-# Inline termStaticCheck #-}